Make publishing stream events wait for the message to be published
To allow message publishing in a ordered way wait for the message
publishing to be ACK by the broker before publishing next one. This
ensure sending messages in an ordered way even if the underlying
broker implementation does not supports message ordering.
Waiting for ACK may introduce latency, because of the broker
availability. To avoid blocking the main thread(UI or replication)
wait for the ACK in a separate thread.

NOTE: This change does not guarantee message *delivery* ordering but
message producing ordering. Message delivery ordering depends on the
broker implementation and should be addressed in a separate change.

Bug: Issue 14909
Change-Id: I99628f78fd9fe60bfadf3ab913d47bb9280727ae