Update maven-javadoc-plugin for Java 11+
The javadoc goal fails with Java 9 or later when JAVA_HOME is not set,
the plugin is unable to find the javadoc executable. The reason is
"javadoc" has been moved.

The issue has been solved in version 3.3.0 of the plugin:

https://issues.apache.org/jira/browse/MJAVADOC-650

Update the maven-javadoc-plugin to latest release 3.4.0.

I have done a similar update at:
https://github.com/victools/jsonschema-generator/pull/262

Also remove the `source=8` javadoc configuration for
https://bugs.openjdk.org/browse/JDK-8212233 The pom targets Java 11.

Change-Id: If04cbe2dbdbf9583f154f7c9e1e448e74df2c19d