Add interface ExtendedBrokerApi to consume messages with groupId
The intention of this change is to extend the current API allowing
consumers read meesages with the use of a groupId, which  enables
multiple consumers to independently read messages from the same source,
facilitating parallelism, fault tolerance, and efficient processing.
It allows for workload distribution and ordered processing within
consumer groups.

Change-Id: I3404d930592cd62377181062e7eb474ddd3e5dc7