Provide subscribers with consumer's group id
This change extends the ExtendedBrokerApi interface with the
capability to provide the list of subscribers with their
consumer's group id.

The class InProcessBrokerApi has been updated to implement the
ExtendedBrokerApi interface instead of BrokerApi, it's crucial to
emphasize that the internal message broker lacks support for the
consumer's group id concept. As a result, the logic does not
incorporate this feature.

It's worth noting that libraries using this interface may implement
it if their underlying broker technology supports the concept of
consumer group identifiers.

Bug: Issue 299327285
Change-Id: I73652f69ee184a3247dd104a0a6b353da3fef89d