Provide subscribers with consumer's group id
The intention of this change is to extend the ExtendedBrokerApi
interface with the capability to provide the list of subscribers
with their consumer's group id.

Now the class InProcessBrokerApi implements the ExtendedBrokerApi
interface instead of BrokerApi in order to provide support to
the concept of consumer's group id.

It's worth noting that libraries using this interface may implement
it if their underlying broker technology supports the concept of
consumer group identifiers.

Bug: Issue 299327285
Change-Id: I73652f69ee184a3247dd104a0a6b353da3fef89d