Remove actual implementation of InProcessBrokerApi
The InProcessBrokerApi main purpose was to allow to register
the events listeners independently from the actual implementation
of the broker-api.

The implementation went unfortunately well beyond the initial scope
and started to simulate a full in-process broker implementation
which would not make sense to use in production.
Also, the fact of silently mocking the real behaviour was potentially
leaving configuration and setup issues unnoticed with the drawback
of not notifing the Gerrit admins that something wasn't quite right.

Throw an exception of unsupported method for all of those calls
that are not supposed to be used as actual implementation of a
broker.

Change-Id: I92100d3d447aba32af58cd05e065affb402b7372