[DO NOT MERGE] Provide subscribers with consumer's group id
Extend the BrokerApi interface with the capability to subscribe
to a topic with a specific consumer's group id.

The class InProcessBrokerApi has been updated to implement the
ExtendedBrokerApi interface instead of BrokerApi.

Libraries using this interface may optionally implement it if
their underlying broker technology supports the concept of
consumer group identifiers.

Original-Author: alvaro.vilaplana@gmail.com
Bug: Issue 299327285
Change-Id: I13c1ef0f12161cb06ce01d0891f2f8c09c43caac