Make events-broker a Gerrit plugin
Change the Bazel BUILD style from library to plugin for simplifying
the joint development with other dependent plugins and its import
as an Eclipse project.

Also enable the ability to build other plugins that depend on
events-broker (e.g. events-kafka) by just referring to the source
code and without the requirement to fetch the artifact from Maven.

Change-Id: I86f3ca5f65a15a7e104ced28b44e370424c49a5e