Define a test harness for the implementation of BrokerApi
Make the BrokerApiTest an abstract class that is meant to be used as
test harness for all implementations of the BrokerApi.

E.g. The KafkaBrokerApi would have then a subclass of the BrokerApiTest
and will need to implement only the brokerApi() method to create the
instance of the BrokerApi that is under test.

Change-Id: I0fc789580962a328468b3752647ff44cc479ae4d