Define a test harness for the implementation of ExtendedBrokerApi
Add  ExtendedBrokerApiTest an abstract class, meant to be used as
test harness for all implementations of the BrokerApi.

E.g. The KafkaBrokerApi would have then a subclass of the
ExtendedBrokerApiTest and will need to implement only the
brokerApi() method to create the instance of the BrokerApi that is
under test.

Change-Id: I97c5ec498c8a144c8f59db91eb9ef618e2e53009