Move message logging to events broker
The message log file is currently populated in an inconsistent way when
it comes to publishing messages. Publishing of some messages, like index
events, are still logged, however anything related to the stream events
topic is not, since this functionality was moved from multi-site to
events-broker. But it looks like the logging aspect was left out at the
time.

It makes sense to move all the message logging classes to the `events-
broker`, so they are together with the logic that publishes the stream
events, and used directly. This way, the clients of the library will not
even need to worry (or care) about updating the message log, this will
happen transparently.

The logic in `StreamEventPublisher` was already dealing with updating
the metrics. Incrementing the metrics is another feature that should be
tightly coupled with the publishing of the messages, however it looks
like some changes are required to allow each client of the library to
inject their own broker metrics. This will be done in a separate change.

Bug: Issue 294904654
Change-Id: I4e27619a0dde52c29626a33d2bcfb93d5945d509