Move message logging to events broker
The message log file is currently populated in an inconsistent way when
it comes to publishing messages. Publishing of some messages, like index
events, are still logged, however anything related to the stream events
topic is not, since this functionality was moved from multi-site to
events-broker. But it looks like the logging aspect was forgotten.

It makes sense to move all the message logging classes to the `events-
broker`, as they will then be available to all the publishers which
depend on this library, like `multi-site`, `events-kafka`,
`events-aws-kinesis` etc.

Bug: Issue 294904654
Change-Id: I4e27619a0dde52c29626a33d2bcfb93d5945d509