Move BrokerApiWrapper from multi-site to events-broker
This change completes the work to segregate updating of the message_log
file when an event is published via the broker.

multi-site's `BrokerApiWrapper` class is moved into `events-broker`,
but without the metric publishing, which will be done in a separate
change. As such, for now the updating of the metrics is the client's
responsibility, something that is reflected in `StreamEventPublisher`
and will also be reflected in multi-site plugin.

It is important to note that an EventsBrokerExecutor replaces the
StreamEventsBrokerExecutor for the publishing of the stream events.
However, the `EventsBrokerExecutor` can be used more broadly for the
publishing of any event, not just stream events. In addition, most of
the StreamEventPublisher logic is now moved to the new wrapper class.

Another important change is that the logic over whether to publish
the stream event is slightly different; before, the event would still
be published if both the instance id of the node and the instance id of
the message were null. This is something that was also enforced in the
tests.  However, the logic from multi-site was slightly different; when
both instanceIds are null, the message is skipped. This was something
that was flagged during the kick off and the direction given was that
the multi-site logic is safe to be reused across the events-broker.

Bug: Issue 294904654
Change-Id: Ic39a58a2d0916ee26dfb3a28f73713cc88acac38