Move BrokerApiWrapper from multi-site to events-broker
multi-site's `BrokerApiWrapper` class is moved into `events-broker`
`EventsBrokerApiWrapper`, but without the metric publishing, which will
be done in a separate change. As such, for now the updating of the
metrics becomes the client's responsibility, something that is reflected
in multi-site plugin. In addition, the logic that decides when an event
should be skipped from publishing also remains in multi-site.

In theory, event-broker's `StreamEventPublisher` can also be adapted to
use the new wrapper class, but this can be done in a follow up change.

Bug: Issue 294904654
Change-Id: Ic39a58a2d0916ee26dfb3a28f73713cc88acac38