Move BrokerApiWrapper from multi-site to events-broker
This change completes the work to segregate updating of the message_log
file when an event is published via the broker.

multi-site's `BrokerApiWrapper` class is moved into `events-broker`,
but without the metric publishing, which will be done in a separate
change. As such, for now the updating of the metrics is the client's
responsibility, something that is reflected in `StreamEventPublisher`
and will also be reflected in multi-site plugin.

Bug: Issue 294904654
Change-Id: Ic39a58a2d0916ee26dfb3a28f73713cc88acac38