Introduce new interface method for disconnecting consumers by groupId
It is useful to allow plugins to isolate their consumers using the
groupId, so that multiple consumers can work independently.

Allow the broker API to disconnect only specific consumers on
topics by groupId, so that unloading one plugin would not impact
the functionality of other plugins consuming messages on the
same topics.

Change-Id: I859cddd012295e576fc4df4a66af65eedfd24d29